Change BIGINT to INTEGER

+++ src/data/migrations/20180930204039-create-email-activation-code.js
@@ -14,7 +14,7 @@ module.exports = {
         field: 'activation_code'
       },
       expirationTime: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         field: 'expiration_time'
       },
       userId: {
